# String#squeeze fastpath single char case