@chrisseaton 2021-05-17
Fascinating method! And I see the code path sort of unifies with `tr`.